Start shared package extraction and mobile scaffold
