+++ mobile/README.md
@@ -1,0 +1,5 @@
+# Mobile app scaffold
+
+This directory is reserved for the upcoming Expo React Native app.
+
+The initial scaffold demonstrates that mobile code can import platform-safe modules from `packages/shared`.
