+++ mobile/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "expo/tsconfig.base",
+  "compilerOptions": {
+    "strict": true
+  },
+  "include": ["./**/*.ts", "./**/*.tsx", "../packages/shared/src/**/*.ts"]
+}
